+++ src-tauri/src/utils/mod.rs
@@ -8,7 +8,6 @@ pub mod heic_decode;
 pub mod orientation;
 pub mod raw_decode;
 pub mod raw_file;
-pub mod video_probe;
 
 // Re-export commonly used cache functions for backwards compatibility
 pub use cache::{generate_cache_path, generate_persistent_cache_path, get_cache_dir};
